requirements.txt
fastapi
uvicorn
celery
redis
networkx
gevent
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==3.2.2
scikit-learn
neo4j
flwr
pytest
httpx
